Tweak README, add interfaces/types for data models
This commit adds TypeScript interfaces to represent the models for data the application will work with. I like starting with these as it makes building the application simpler when you know what data different parts are passing around.

A rough outline:

On the payment gateway side, there's an interface for the response it should send to our API - a number ID for the transaction, a boolean to track whetherthe payment occurred, and an error property that's either null or a constant noting insufficient funds. I went with a constant to make it easier to add oth
er constants down the line, perhaps if other types of errors needed to be supported.

The API contains an interface for the data actually coming from the customer: first and last name, credit card number, expiration month and expiration year. (We'd want to add CVC code at some point, too). Numbers in JavaScript support a maximum of 16 digits and some Googling tells me credit cards in some countries get up to 19 digits. So I went with a string credit card number that I plan to regex and make sure it contains nothing but digits. Would be curious to see how your team handles this issue.

There's also an interface for the response the API sends to clients: either an error or data object, each with string message properties explaining what happened.

Finally, there's a small tweak to the todo list based on something Nathan mentioned to me regarding retries of the payment gateway.