Add payment gateway handling for bad/missing Authorization header
This commit begins work on the mock payment gateway. The gateway first checks all responses for the presence of an Authorization HTTP header. If the header is missing, or the basic auth doesn't match the username and pass in the spec, the gateway returns a 401 Unauthorized response.

The commit includes a couple of tests using the Mocha framework and chai-HTTP, an awesome package that makes it easy to ping your own Express server with various requests and asser against the response.

Also included: minor tweaks to Typescript + NPM configuration and the inclusion of dotenv, which lets us read the required auth credentials from a git ignored environment variable file.