Finish implementation of mock payment gateway
This commit completes the mock payment gateway, which now handles requests with valid Authorization headers. The gateway first generates a random intger between 0 and 2. Based on the number, it either
- Sends a 200 OK header and an object with a random string ID, 'paid' set to true and a null error
- Sends a 200 OK header and and object with a random string ID, 'paid' set to false and an error message set to 'INSUFFICIENT_FUNDS', or
- Sends a 503 Service Unavailable header and an object with a random String ID, 'paid' set to false and an error message set to 'SERVICE_UNAVAILABLE'

The commit includes tests to force the gateway down each of these paths and assert its response.

Other minor things:

- A new utility file with a function to generate a random integer given a max (used for the payment gateway's random nature and stolen shamelessly from MDN)
- Another utility function to generate a random string using Node's crypto module
- A change to the CustomerData interface - JS numbers can't start with zero, so we have to use a string to represent a customer's CC expiration date. Now I'm really curious to see how the team handles this stuff.

I'll be merging this back into the master branch to start work on the customer-facing API.