Complete customer-facing API code and tests (save for database persistence)
This commit completes the customer-facing API's possible responses based on user data and the payment gateway's response. The previous commit covered the happy path, when the payment gateway returns 'sufficient funds' on the first ping. This commit adds the following scenarios:

- The payment gateway returns 'invalid credentials' on the first ping. In this case, the API returns a 400 Bad Request to the client.
- The payment gateway returns 'insufficient funds' on the first ping. Here the API also returns a 400 Bad Request - though I'm curious as to whether the team thinks that's the best possible response.

This commit also handles when the payment gateway returns a 503 unavailable status code. The HTTP package I selected throws an error in that case - so we check the error to see if it's in fact a 503. If it is, we return null instead.

That null response tells the API to re-ping the payment gateway up to two more times. If it gets a response in one of those tries - sufficient payment, insufficient, bad credentials, whatever - it handles them like normal. If, after three tries, it's seen nothing but 503s, it gives up and sends the client a 503 instead.

The test suite includes a fair number of Nock fake server calls to trigger the appropriate number of 503s before giving a valid payment gateway response to force our code down these various paths.

This commit also includes a new constant for the message the payment gateway sends upon receiving bad credentials.