Add mock database functionality
This commit adds a very basic in-memory mock database using a JavaScript closure. The DB supports methods to insert a row of customer data (it adds a Date subscriptionDate property), select every 'row' and clear the database. The customer-facing API uses this mock database to 'insert' a subscriber if they successfully pay.